+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Changes to Calva.
 
 ## [Unreleased]
+- [Fix colors in suggestion popup (REPL window)](https://github.com/BetterThanTomorrow/calva/issues/623)
+
+## [2.0.95] - 2020-04-25
+- [Separate setting for highlighting current indent guide](https://github.com/BetterThanTomorrow/calva/issues/625)
+- [Fix: Problems with v2.0.94 rendering performance ](https://github.com/BetterThanTomorrow/calva/issues/626)
+
+## [2.0.94] - 2020-04-24
+- [Rainbow indent guides](https://github.com/BetterThanTomorrow/calva/issues/620)
 
 ## [2.0.93] - 2020-04-21
 - [Unclutter editor context menu when not in clojure files](https://github.com/BetterThanTomorrow/calva/issues/615)